details is showing data from db now

+++ app/src/main/java/com/denysque/testdemo/core/repository/DatabaseRepository.java
@@ -10,7 +10,7 @@ public interface DatabaseRepository {
 
     List<Forecast> getAllForecasts();
 
-    Forecast getForecastById(int id);
+    Forecast getForecastById(long id);
 
     List<Long> getAllCitiesID();
 }
